Update typeshed to 50d98acc766e9425cb099c17b711f1bccb697584.
Also updates pytype/mixin.py and some test cases to reflect that in py3,
unittest.TestCase is in unittest/case.pyi rather than unittest/__init__.pyi.
