+++ pytype/tests/test_classes.py
@@ -1367,43 +1367,6 @@ class ClassesTest(test_base.TargetIndependentTest):
         x = ...  # type: int
     """)
 
-  def testInitTestClassInSetup(self):
-    ty = self.Infer("""\
-      import unittest
-      class A(unittest.TestCase):
-        def setUp(self):
-          self.x = 10
-        def fooTest(self):
-          return self.x
-    """)
-    self.assertTypesMatchPytd(ty, """
-      import unittest
-      unittest = ...  # type: module
-      class A(unittest.TestCase):
-          x = ...  # type: int
-          def fooTest(self) -> int: ...
-    """)
-
-  def testInitInheritedTestClassInSetup(self):
-    ty = self.Infer("""\
-      import unittest
-      class A(unittest.TestCase):
-        def setUp(self):
-          self.x = 10
-      class B(A):
-        def fooTest(self):
-          return self.x
-    """)
-    self.assertTypesMatchPytd(ty, """
-      import unittest
-      unittest = ...  # type: module
-      class A(unittest.TestCase):
-          x = ...  # type: int
-      class B(A):
-          x = ...  # type: int
-          def fooTest(self) -> int: ...
-    """)
-
   def testPyiNestedClass(self):
     # Test that pytype can look up a pyi nested class in a py file and reconsume
     # the inferred pyi.
